+++ ChangeLog.markdown
@@ -8,6 +8,7 @@
 * Allow selecting date and time (without closing filter) in date filtering https://github.com/vedmack/yadcf/issues/296
 * Added global property - filters_tr_index: Allow to control the index of the <tr> inside the thead of the table, e.g when one <tr> is used for headers/sort and another <tr> is used for filters https://github.com/vedmack/yadcf/issues/297
 * Fixed values with quotation marks filtered incorrectly by filter_type select https://github.com/vedmack/yadcf/issues/317
+* Added moment_date_format for better parsing when using datepicker_type: 'bootstrap-datetimepicker'
 * Some of additional closed issues https://github.com/vedmack/yadcf/issues/295 / https://github.com/vedmack/yadcf/issues/308 / https://github.com/vedmack/yadcf/issues/314