+++ ChangeLog.markdown
@@ -1,6 +1,11 @@
 # Yet Another DataTables Column Filter (yadcf) Change-log
 
-## 0.9.0 still in beta (grab latest stable from https://github.com/vedmack/yadcf/releases)
+## 0.9.1 still in beta (grab latest stable from https://github.com/vedmack/yadcf/releases)
+* Fixed autocomplete in cumulative mode filtering
+
+
+
+## 0.9.0
 
 * Fixed autocomplete with dt in ajax source https://github.com/vedmack/yadcf/issues/282
 * New option, omit_default_label - Prevent yadcf from adding "default_label" (Select value / Select values)